Execution mode changes
In accord with Unix Philosphy, we no longer force interactive mode. Interactive mode must be explicitly called for with option -i or --interactive.

The new default is to process script on standard input if no file has been specified plainly as the first argument or with option -c SCRIPT or --command SCRIPT.

Script on standard input will be read completely to end of file before any evaluation takes place.

It was fun to run like bash does normally, but that is not the purpose of XPR or Krypton. Interactive mode is with the project for mainly for debugging purposes.

While i would like to keep the old way, java does not provide any way to determine what type of stream is being read on standard input, so we have to make this determination via command options. Furthermore, JavaScript is not really a console friendly Subscript.

Unix Philosphy Wins.