Shamelessly Ripped Mozilla's SourceReader tool.
Enhanced script loading features. Technically, this is not a modification of Rhino Source, but a conglomeration of a few rhino sources, with an addition which will deny reading of remote content if the allowRemote flag is set. This code does not live in the Mozilla Source package, but in git.hsusa.core.io, and it has no Mozilla Source Tree dependencies.

Additionaly, git.hsusa.core.io is a public Namespace of the hscore project, so if anyone wants the source, you should look there..