Bash Based Unix Installer
Nixed wrong usage of builtin variable.

Notes:

Because of the fact that this installer uses, bash, `bash' is a
requirement for all unix environments.

JVM 1.8 is the only known external Krypton runtime-requirement at this
time.

Krypton doesn't have any external requirements. Everything is provided
within the public release (this) repository.